chore: cleanup v2
